Fix warning about GtkHTMLSaveReceiverFn
```
main.c: 在函数‘view_source_dialog’中:
main.c:207:31: 警告：cast between incompatible function types from ‘gboolean (*)(HTMLEngine *, const gchar *, guint,  GString *)’ {或称 ‘int (*)(struct _HTMLEngine *, const char *, unsigned int,  struct _GString *)’} to ‘gboolean (*)(void *, const gchar *, gsize,  void *)’ {或称 ‘int (*)(void *, const char *, long unsigned int,  void *)’} [-Wcast-function-type]
  207 |                 content_type, (GtkHTMLSaveReceiverFn)
      |                               ^
gtkhtml-editor.c: In function 'gtkhtml_editor_get_text_html':
gtkhtml-editor.c:1371:17: warning: cast between incompatible function types from 'gboolean (*)(HTMLEngine *, const gchar *, guint,  GString *)' {aka 'int (*)(struct _HTMLEngine *, const char *, unsigned int,  struct _GString *)'} to 'gboolean (*)(void *, const gchar *, gsize,  void *)' {aka 'int (*)(void *, const char *, long unsigned int,  void *)'} [-Wcast-function-type]
 1371 |                 (GtkHTMLSaveReceiverFn) editor_save_receiver, string);
      |                 ^
gtkhtml-editor.c: In function 'gtkhtml_editor_get_text_plain':
gtkhtml-editor.c:1391:17: warning: cast between incompatible function types from 'gboolean (*)(HTMLEngine *, const gchar *, guint,  GString *)' {aka 'int (*)(struct _HTMLEngine *, const char *, unsigned int,  struct _GString *)'} to 'gboolean (*)(void *, const gchar *, gsize,  void *)' {aka 'int (*)(void *, const char *, long unsigned int,  void *)'} [-Wcast-function-type]
 1391 |                 (GtkHTMLSaveReceiverFn) editor_save_receiver, string);
      |                 ^
```

+++ components/editor/gtkhtml-editor.c
@@ -1348,7 +1348,7 @@ gtkhtml_editor_set_spell_languages (GtkhtmlEditor *editor,
 static gboolean
 editor_save_receiver (HTMLEngine *engine,
                       const gchar *data,
-                      guint length,
+                      gsize length,
                       GString *contents)
 {
 	g_string_append_len (contents, data, length);
